requirements.txt
asgiref==3.4.1
Django==4.0.1
numpy==1.21.5
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
sqlparse==0.4.2
tzdata==2021.5
pytest-django==4.5.2